--- a/src/Esprima/Ast/SyntaxToken.cs
+++ b/src/Esprima/Ast/SyntaxToken.cs
@@ -15,7 +15,13 @@ public sealed class SyntaxToken : SyntaxElement
     public TokenType Type { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
 
     public string Value { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
-    public RegexValue? RegexValue { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
+    public RegexValue? RegexValue
+    {
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        get => (RegexValue?) _additionalDataContainer.InternalData;
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private set => _additionalDataContainer.InternalData = value;
+    }
 
     public override string ToString() => Value;
 }