… src/Esprima/Token.cs
… test/Esprima.Tests/LocationTests.cs