Besides removing JSX-awareness from core parser, this PR also proposes moving `ParsedToken`/`ParsedComment` into the `Esprima.Ast` namespace and changing them into node-like types because the information contained by these types belongs to the AST (see e.g. the tree displayed in AST Explorer). Roslyn, for example, models tokens and comments similarly.