@lahma 2022-08-15
## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|10.948 ms|0.0251 ms|421.8750|203.1250|7,004 KB|
| **New** |	|	| **10.454 ms (-5%)** | **0.0233 ms** | **406.2500 (-4%)** | **156.2500 (-23%)** | **6,892 KB (-2%)** |
| Old |ParseProgram|backbone-1.1.0|1.406 ms|0.0098 ms|64.4531|31.2500|1,055 KB|
| **New** |	|	| **1.376 ms (-2%)** | **0.0040 ms** | **62.5000 (-3%)** | **25.3906 (-19%)** | **1,036 KB (-2%)** |
| Old |ParseProgram|jquery-1.9.1|8.381 ms|0.0200 ms|359.3750|171.8750|6,086 KB|
| **New** |	|	| **8.706 ms (+4%)** | **0.0343 ms** | **359.3750 (0%)** | **171.8750 (0%)** | **5,976 KB (-2%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|13.532 ms|0.0368 ms|546.8750|265.6250|9,108 KB|
| **New** |	|	| **13.337 ms (-1%)** | **0.0272 ms** | **531.2500 (-3%)** | **265.6250 (0%)** | **8,901 KB (-2%)** |
| Old |ParseProgram|mootools-1.4.5|6.851 ms|0.0134 ms|296.8750|101.5625|4,901 KB|
| **New** |	|	| **6.632 ms (-3%)** | **0.0151 ms** | **289.0625 (-3%)** | **109.3750 (+8%)** | **4,812 KB (-2%)** |
| Old |ParseProgram|underscore-1.5.2|1.199 ms|0.0023 ms|58.5938|23.4375|959 KB|
| **New** |	|	| **1.200 ms (0%)** | **0.0020 ms** | **56.6406 (-3%)** | **19.5313 (-17%)** | **943 KB (-2%)** |
| Old |ParseProgram|yui-3.12.0|6.071 ms|0.0111 ms|265.6250|117.1875|4,403 KB|
| **New** |	|	| **6.335 ms (+4%)** | **0.0148 ms** | **257.8125 (-3%)** | **125.0000 (+7%)** | **4,301 KB (-2%)** |