@adams85 2022-08-15
Added a further commit (https://github.com/sebastienros/esprima-dotnet/pull/310/commits/8d8b0b398cc5a43831c5b214825672ce32d10d50) for the case where `ParserOptions.Token` is enabled. This keeps the memory footprint of tokens the same as before (by utilizing the otherwise unused internal additional data slot).

@lahma So, do you agree with the proposed changes, especially with https://github.com/sebastienros/esprima-dotnet/pull/310/commits/d09a6c76ee82651c1dc596a9fa87b9c16241a5e0?