@lahma 2022-08-15
> Added a further commit ([8d8b0b3](https://github.com/sebastienros/esprima-dotnet/commit/8d8b0b398cc5a43831c5b214825672ce32d10d50)) for the case where `ParserOptions.Token` is enabled. This keeps the memory footprint of tokens the same as before (by utilizing the otherwise unused internal additional data slot).
> 
> @lahma So, do you agree with the proposed changes, especially with [d09a6c7](https://github.com/sebastienros/esprima-dotnet/commit/d09a6c76ee82651c1dc596a9fa87b9c16241a5e0)?

I think it looks good, we can always re-iterate when needed 👍🏻